pyproject.toml
[project]
name = "doc-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "fastapi[all]>=0.115.11",
    "loguru>=0.7.3",
    "markitdown[docx,pdf,pptx,xls,xlsx]>=0.1.0",
    "openai>=1.68.2",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "pymupdf>=1.25.4",
    "python-magic>=0.4.27",
    "qwen-vl-utils>=0.0.10",
    "torchvision>=0.21.0",
]
